adds MozillaTaskbarPreviewClass float rule
Adds a float rule for windows with class `MozillaTaskbarPreviewClass`.  These invisible windows are spawned by Firefox when its "Show tab previews in the Windows taskbar" option is enabled, presumably as a way to populate thumbnails for inactive tabs.  Komorebi will then try to tile these and make a mess of the layout.