fix(drag): obsidian + discord no longer need force
This commit removes the force option from Obsidian and Discord, which no
longer need to be force-managed as of v0.1.22. With force enabled, when
dragging items onto these apps, a helper window tied to the same Exe
with the title "Drag" will be force-managed alongside the actual
application window, which results in significant ghosting.